Revert "Ooops"
This reverts commit deccf88124018f1717d7f064392b653252a74abd.

+++ modules/gin_simd/mipp/mipp_impl_SSE.hxx
@@ -3564,7 +3564,7 @@
     }
 
     template <>
-    inline reg trunc<double>(const reg v) {
+    inline reg trunc<float>(const reg v) {
         return _mm_cvtepi32_pd (_mm_cvttpd_epi32 (v));
     }
 #endif
